rm globals module
