--- Sources/Globals/StringError.swift
@@ -1,1 +1,0 @@
-extension String: Error {}
