+++ Sources/VaporToolbox/New/New.swift
@@ -1,4 +1,3 @@
-import Globals
 import ConsoleKit
 import Foundation
 import Yams
@@ -14,9 +13,6 @@ struct New: Command {
     func run(using ctx: CommandContext, signature: Signature) throws {
         let name = signature.name
         let gitUrl = "https://github.com/vapor/template"
-        // TODO: allow for dynamic template once format is documented
-//        let template = signature.expandedTemplate()
-//        let gitUrl = try template.fullUrl()
 
         // Cloning
         ctx.console.info("Cloning template...")
@@ -71,55 +67,6 @@ struct New: Command {
     }
 
 }
-
-//extension New.Signature {
-//    func expandedTemplate() -> Template {
-//        guard let chosen = self.template else { return .default }
-//        switch chosen {
-//        case "web": return .web
-//        case "api": return .api
-//        case "auth": return .auth
-//        default: return .custom(repo: chosen)
-//        }
-//    }
-//}
-//
-//
-//enum Template {
-//    case `default`, web, api, auth
-//    case custom(repo: String)
-//
-//    fileprivate func fullUrl() throws -> String {
-//        switch self {
-//        case .default: return "https://github.com/vapor/template"
-//        case .api: return "https://github.com/vapor/api-template"
-//        case .web: return "https://github.com/vapor/web-template"
-//        case .auth: return "https://github.com/vapor/auth-template"
-//        case .custom(let custom): return try expand(templateUrl: custom)
-//        }
-//    }
-//
-//
-//    /// http(s)://whatever.com/foo/bar => http(s)://whatever.com/foo/bar
-//    /// foo/some-template => https://github.com/foo/some-template
-//    /// some-template => https://github.com/vapor/some-template
-//    /// some => https://github.com/vapor/some
-//    /// if fails, attempts `-template` suffix
-//    /// some => https://github.com/vapor/some-template
-//    private func expand(templateUrl url: String) throws -> String {
-//        // all ssh urls are custom
-//        if url.contains("@") { return url }
-//
-//        // expand github urls, ie: `repo-owner/name-of-repo`
-//        // becomes `https://github.com/repo-owner/name-of-repo`
-//        let components = url.split(separator: "/")
-//        if components.count == 1 { throw "unexpected format, use `repo-owner/name-of-repo`" }
-//
-//        // if not 2, then it's a full https url
-//        guard components.count == 2 else { return url }
-//        return "https://github.com/\(url)"
-//    }
-//}
 
 struct PrintDroplet: Command {
     struct Signature: CommandSignature {}
