+++ Sources/VaporToolbox/Supervisor/SupervisorUpdate.swift
@@ -12,6 +12,6 @@ struct SupervisorUpdate: Command {
     func run(using context: CommandContext, signature: Signature) throws {
         let package = try Process.swift.package.dump()
         try Process.run(Shell.default.which("supervisorctl"), "update", package.name)
-        context.console.info("Supervisor entry for \(package.name) updated")
+        context.console.info("Supervisor entry for \(package.name) updated.")
     }
 }
